Organize modules into subpackages
