+++ analyzer/email_clients/__init__.py
@@ -1,0 +1,1 @@
+"""Email client connectors."""
