+++ analyzer/reporting/__init__.py
@@ -1,0 +1,1 @@
+"""Reporting utilities."""
