+++ analyzer/threat/__init__.py
@@ -1,0 +1,1 @@
+"""Threat analysis helpers."""
