+++ analyzer/threat/traffic_light.py
