+++ tests/test_email_scanner.py
@@ -2,6 +2,7 @@
 """Tests for the e-mail scanner module."""
 
 import sys
+import types
 from unittest.mock import MagicMock, patch
 
 # Externe Abhängigkeiten mocken, damit Imports in email_scanner nicht fehlschlagen
@@ -20,19 +21,20 @@ sys.modules["requests"] = MagicMock()
 sys.modules["colorama"] = MagicMock()
 
 # TrafficLight während des Imports stubben und anschließend zurücksetzen
-original_traffic_light = sys.modules.get("analyzer.traffic_light")
+original_traffic_light = sys.modules.get("analyzer.threat.traffic_light")
 mock_traffic_light = MagicMock()
 mock_traffic_light.analyze_threat_level = MagicMock()
-sys.modules["analyzer.traffic_light"] = mock_traffic_light
+sys.modules.setdefault("analyzer.threat", types.ModuleType("analyzer.threat"))
+sys.modules["analyzer.threat.traffic_light"] = mock_traffic_light
 
 import analyzer.email_scanner  # noqa: E402
 from analyzer.email_scanner import get_outlook_emails, scan_inbox  # noqa: E402
 
 # Ursprüngliches TrafficLight-Modul wiederherstellen
 if original_traffic_light is not None:
-    sys.modules["analyzer.traffic_light"] = original_traffic_light
+    sys.modules["analyzer.threat.traffic_light"] = original_traffic_light
 else:  # pragma: no cover - falls Modul zuvor nicht existierte
-    del sys.modules["analyzer.traffic_light"]
+    del sys.modules["analyzer.threat.traffic_light"]
 
 
 def test_get_outlook_emails_dummy_scanner():
