+++ tests/test_local_ai_handler.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from analyzer.local_ai_handler import LocalAIHandler
+from analyzer.ml.local_ai_handler import LocalAIHandler
 
 
 def test_analyze_email_content_combines_results(monkeypatch):
